replace ErrorResponseWithStatus to pipeline/pkg/common
