+++ api/security.go
@@ -27,10 +27,9 @@ import (
 
 	"github.com/banzaicloud/anchore-image-validator/pkg/apis/security/v1alpha1"
 	clientV1alpha1 "github.com/banzaicloud/anchore-image-validator/pkg/clientset/v1alpha1"
-	"github.com/banzaicloud/pipeline/api/ark/common"
 	"github.com/banzaicloud/pipeline/helm"
 	"github.com/banzaicloud/pipeline/internal/security"
-	pkgCommmon "github.com/banzaicloud/pipeline/pkg/common"
+	"github.com/banzaicloud/pipeline/pkg/common"
 	"github.com/banzaicloud/pipeline/pkg/k8sclient"
 	"github.com/banzaicloud/pipeline/pkg/security"
 	"github.com/gin-gonic/gin"
@@ -54,7 +53,7 @@ func getSecurityClient(c *gin.Context) *clientV1alpha1.SecurityV1Alpha1Client {
 	config, err := k8sclient.NewClientConfig(kubeConfig)
 	if err != nil {
 		log.Errorf("Error getting K8s config: %s", err.Error())
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error getting K8s config",
 			Error:   err.Error(),
@@ -65,7 +64,7 @@ func getSecurityClient(c *gin.Context) *clientV1alpha1.SecurityV1Alpha1Client {
 	securityClientSet, err := clientV1alpha1.SecurityConfig(config)
 	if err != nil {
 		log.Errorf("Error getting SecurityClient: %s", err.Error())
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error getting SecurityClient",
 			Error:   err.Error(),
@@ -87,7 +86,7 @@ func GetScanLog(c *gin.Context) {
 	if err != nil {
 		err := errors.Wrap(err, "Error during request processing")
 		log.Error(err.Error())
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error getting scanlogs",
 			Error:   err.Error(),
@@ -124,7 +123,7 @@ func GetWhiteLists(c *gin.Context) {
 	if err != nil {
 		err := errors.Wrap(err, "Error during request processing")
 		log.Error(err.Error())
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error getting whitelists",
 			Error:   err.Error(),
@@ -158,7 +157,7 @@ func CreateWhiteList(c *gin.Context) {
 	if err != nil {
 		err := errors.Wrap(err, "Error parsing request:")
 		log.Error(err.Error())
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error during parsing request!",
 			Error:   errors.Cause(err).Error(),
@@ -183,7 +182,7 @@ func CreateWhiteList(c *gin.Context) {
 	if err != nil {
 		err := errors.Wrap(err, "Error during request processing")
 		log.Error(err.Error())
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error creating whitelist",
 			Error:   err.Error(),
@@ -197,7 +196,7 @@ func CreateWhiteList(c *gin.Context) {
 func DeleteWhiteList(c *gin.Context) {
 	name := c.Param("name")
 	if len(name) == 0 {
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "WhiteList name is required!",
 			Error:   "WhiteList name is required!",
@@ -214,7 +213,7 @@ func DeleteWhiteList(c *gin.Context) {
 	if err != nil {
 		err := errors.Wrap(err, "Error during request processing")
 		log.Error(err.Error())
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error deleting whitelist",
 			Error:   err.Error(),
@@ -230,7 +229,7 @@ func createResponse(c *gin.Context, response http.Response) {
 	err := json.NewDecoder(response.Body).Decode(&responsePayload)
 	if err != nil {
 		log.Error("Error parsing response: %v", err.Error())
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error parsing response",
 			Error:   err.Error(),
@@ -269,7 +268,7 @@ func GetPolicies(c *gin.Context) {
 	response, err := anchore.DoAnchoreRequest(anchoreRequest)
 	if err != nil {
 		log.Error(err)
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error",
 			Error:   err.Error(),
@@ -289,7 +288,7 @@ func CreatePolicy(c *gin.Context) {
 	if err != nil {
 		err := errors.Wrap(err, "Error parsing request:")
 		log.Error(err.Error())
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error during parsing request!",
 			Error:   errors.Cause(err).Error(),
@@ -317,7 +316,7 @@ func CreatePolicy(c *gin.Context) {
 	response, err := anchore.DoAnchoreRequest(anchoreRequest)
 	if err != nil {
 		log.Error(err)
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error",
 			Error:   err.Error(),
@@ -334,7 +333,7 @@ func UpdatePolicies(c *gin.Context) {
 
 	policyId := c.Param("policyId")
 	if len(policyId) == 0 {
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "policyId is required!",
 			Error:   "policyId is required!",
@@ -365,7 +364,7 @@ func UpdatePolicies(c *gin.Context) {
 		response, err := anchore.DoAnchoreRequest(anchoreRequest)
 		if err != nil {
 			log.Error(err)
-			c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+			c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 				Code:    http.StatusInternalServerError,
 				Message: "Error",
 				Error:   err.Error(),
@@ -385,7 +384,7 @@ func UpdatePolicies(c *gin.Context) {
 		if err != nil {
 			err := errors.Wrap(err, "Error parsing request:")
 			log.Error(err.Error())
-			c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+			c.JSON(http.StatusBadRequest, common.ErrorResponse{
 				Code:    http.StatusBadRequest,
 				Message: "Error during parsing request!",
 				Error:   errors.Cause(err).Error(),
@@ -398,7 +397,7 @@ func UpdatePolicies(c *gin.Context) {
 	response, err := anchore.DoAnchoreRequest(anchoreRequest)
 	if err != nil {
 		log.Error(err)
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error",
 			Error:   err.Error(),
@@ -415,7 +414,7 @@ func DeletePolicy(c *gin.Context) {
 
 	policyId := c.Param("policyId")
 	if len(policyId) == 0 {
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "policyId is required!",
 			Error:   "policyId is required!",
@@ -443,7 +442,7 @@ func DeletePolicy(c *gin.Context) {
 	response, err := anchore.DoAnchoreRequest(anchoreRequest)
 	if err != nil {
 		log.Error(err)
-		c.JSON(http.StatusInternalServerError, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusInternalServerError, common.ErrorResponse{
 			Code:    http.StatusInternalServerError,
 			Message: "Error",
 			Error:   err.Error(),
@@ -465,7 +464,7 @@ func GetImageDeployments(c *gin.Context) {
 	if !re.MatchString(imageDigest) {
 		err := fmt.Errorf("Invalid imageID format: %s", imageDigest)
 		log.Error(err)
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error getting K8s config",
 			Error:   err.Error(),
@@ -483,7 +482,7 @@ func GetImageDeployments(c *gin.Context) {
 	activeReleases, err := helm.ListDeployments(nil, c.Query("tag"), kubeConfig)
 	if err != nil {
 		log.Error("Error listing deployments: ", err.Error())
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error listing deployments",
 			Error:   err.Error(),
@@ -494,7 +493,7 @@ func GetImageDeployments(c *gin.Context) {
 	client, err := k8sclient.NewClientFromKubeConfig(kubeConfig)
 	if err != nil {
 		log.Errorf("Error getting K8s config: %s", err.Error())
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error getting K8s config",
 			Error:   err.Error(),
@@ -505,7 +504,7 @@ func GetImageDeployments(c *gin.Context) {
 	pods, err := listPods(client, "", "")
 	if err != nil {
 		log.Errorf("Error getting pods from cluster: %s", err.Error())
-		c.JSON(http.StatusBadRequest, pkgCommmon.ErrorResponse{
+		c.JSON(http.StatusBadRequest, common.ErrorResponse{
 			Code:    http.StatusBadRequest,
 			Message: "Error getting pods from cluster",
 			Error:   err.Error(),
